Let other users test other certifications authority

+++ images/irods/install_irods.sh
@@ -81,8 +81,10 @@ mkdir -p $certdir
 
 # Copy current certification authority for other containers
 cadir="$certdir/caauth"
-sudo rm -rf $cadir && rsync -avq /etc/grid-security/certificates/ $cadir
-echo "Saved auth for outside containers"
+if [ ! -d "$cadir" ]; then
+    sudo rm -rf $cadir && rsync -avq /etc/grid-security/certificates/ $cadir
+    echo "Saved auth for outside containers"
+fi
 
 # Add a guest user for GSI certificates
 rm -rf /opt/certificates/$GSI_USER && /addusercert $GSI_USER
